working button press
